+++ tedi/components/buttons/collapse-button/collapse-button.stories.ts
@@ -299,9 +299,7 @@ export const SecondaryButton: StoryObj<StatesArgs> = {
 export const WithTextInverted: StoryObj<StatesArgs> = {
   parameters: PSEUDO_PARAMS,
   globals: {
-    backgrounds: {
-      value: "brand",
-    },
+    backgrounds: { value: "brand" },
   },
   args: {
     hideText: false,
@@ -317,9 +315,7 @@ export const WithTextInverted: StoryObj<StatesArgs> = {
 export const IconOnlyInverted: StoryObj<StatesArgs> = {
   parameters: PSEUDO_PARAMS,
   globals: {
-    backgrounds: {
-      value: "brand",
-    },
+    backgrounds: { value: "brand" },
   },
   args: {
     hideText: true,