+++ tedi/components/content/index.ts
@@ -1,4 +1,5 @@
 export * from "./accordion";
+export * from "./card";
 export * from "./list/list.component";
 export * from "./text-group/";
 export * from "./carousel";